benchmark: differentiate whatwg and legacy url
PR-URL: https://github.com/nodejs/node/pull/47377
