+++ benchmark/url/whatwg-url-canParse.js
