+++ benchmark/url/whatwg-url-to-and-from-path.js
